Rename to xmtp_api_http
