+++ .vscode/settings.json
@@ -5,7 +5,7 @@
     "bindings_node/Cargo.toml",
     "bindings_wasm/Cargo.toml",
     "examples/cli/Cargo.toml",
-    "xmtp_api_mls_gateway/Cargo.toml"
+    "xmtp_api_http/Cargo.toml"
   ],
   "rust-analyzer.procMacro.enable": true,
   "rust-analyzer.procMacro.attributes.enable": true,
