+++ xmtp_api_http/Cargo.toml
@@ -1,17 +1,17 @@
 [package]
 edition = "2021"
-name = "xmtp_api_mls_gateway"
+name = "xmtp_api_http"
 version = "0.1.0"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-async-trait = "0.1.68"
-futures = "0.3.29"
+async-trait = { workspace = true }
+futures = { workspace = true }
 reqwest = { version = "0.12.5", features = ["json", "stream"] }
-serde = "1.0"
-serde_json = "1.0"
+serde = { workspace = true }
+serde_json = { workspace = true }
 xmtp_proto = { path = "../xmtp_proto", features = ["proto_full"] }
 
 [dev-dependencies]
