+++ xmtp_api_http/rustfmt.toml
