+++ xmtp_api_http/src/util.rs
@@ -1,0 +1,30 @@
+use serde::de::DeserializeOwned;
+use serde::{Deserialize, Serialize};
+use xmtp_proto::api_client::{Error, ErrorKind};
+
+#[derive(Deserialize, Serialize)]
+#[serde(untagged)]
+enum GrpcResponse<T> {
+  Ok(T),
+  Err(ErrorResponse),
+}
+
+#[derive(Deserialize, Serialize)]
+struct ErrorResponse {
+  code: usize,
+  message: String,
+  details: Vec<String>,
+}
+
+/// handle JSON response from gRPC, returning either
+/// the expected deserialized response object or a gRPC [`Error`]
+pub fn handle_error<S: AsRef<str>, T>(text: S) -> Result<T, Error>
+where
+  T: DeserializeOwned,
+{
+  match serde_json::from_str(text.as_ref()) {
+    Ok(GrpcResponse::Ok(response)) => Ok(response),
+    Ok(GrpcResponse::Err(e)) => Err(Error::new(ErrorKind::IdentityError).with(e.message)),
+    Err(e) => Err(Error::new(ErrorKind::QueryError).with(e.to_string())),
+  }
+}
